Remove 'years active' count from Lobbyist and Org detail views
Marjorie found the 'years active' count to be confusing because it
only considers employer data (which goes back to 2013) rather than
incorporating the first and last contribution/expenditures (which
go back to 2011). In the future we may want to update this count to
consider first and last contributions/expenditures, but for now simply
remove it from the detail views.