Add opening balance and closing balance to filings (#199)
In order to accurately reflect the cash on hand for candidates and committees we need the opening and closing balance from the filings. To do that we need to pull in the data on filings. This PR does that by adding a number of new import scripts that import scraped data about candidates, pacs, and the filings. It also adjust the import script that imports transactions to link those to previously imported filings.

Other changes include moving filing_date attribute from the FilingPeriod class to the Filing class.